protect roles in use from delete

+++ apps/auth/src/role/role.service.ts
@@ -4,10 +4,15 @@ import { RoleRepository } from './role.repository';
 import { RoleDocument } from 'libs/common/schemas/role.schema';
 import { FilterQuery } from 'mongoose';
 import { UpdateRoleInput } from 'libs/common/dtos/inputs/role';
+import { UserService } from '../user/user.service';
+import { RpcException } from '@nestjs/microservices';
 
 @Injectable()
 export class RoleService {
-  constructor(private roleRepository: RoleRepository) {}
+  constructor(
+    private roleRepository: RoleRepository,
+    private userService: UserService,
+  ) {}
 
   async create(roleDTO: any, meta: IMeta) {
     return await this.roleRepository.create({
@@ -34,8 +39,14 @@ export class RoleService {
     return await this.roleRepository.findOneAndUpdate(filterQuery, roleDTO);
   }
 
-  async delete(id: string, meta: IMeta) {
-    const filterQuery: FilterQuery<RoleDocument> = { _id: id };
+  async delete(_id: string, meta: IMeta) {
+    const existUserWithRole = await this.userService.findOneByQuery({
+      role: {
+        $in: [_id],
+      },
+    });
+    if (existUserWithRole) throw new RpcException('Role in use');
+    const filterQuery: FilterQuery<RoleDocument> = { _id };
     if (meta.organization) filterQuery.organization = meta.organization;
     return await this.roleRepository.findOneAndDelete(filterQuery);
   }
